// src/app/modules/profile/model/userProfile.ts
export class UserProfile {
  RoleId: string = '';
  DisplayName:string='';
  RoleName: string = '';
  UserName: string = '';
  AadharNo: string = '';
  Designation: string = '';
  BloodGroup: string = '';
  EmpMobNo: string = '';
  EmpEmail: string = '';
  Password:string='';
  Encryptpwd:string='';
}